--- src/page/NotFound.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function NotFound() {
-  return (
-    <p>404 NOT FOUND</p>
-  );
-}
