+++ src/page/NotFoundPage.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function NotFound() {
+  return (
+    <p>404 NOT FOUND</p>
+  );
+}
