+++ src/page/NotFoundPage.test.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { render } from '@testing-library/react';
+
+import NotFound from './NotFound';
+
+test('NotFound', () => {
+  render(<NotFound />);
+});
